+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.0.40
 
+- Hello, Scrollbar!
 - Default to 'alt' Tk theme on Linux
 - Added Windows/Linux accelerators to Hello, Menu Bar!
 - Fix issue with extra junk character added to beginning of Hello, Checkbutton! by mistake